VCST-5661: Move in-place MapTo methods onto IXCartMapper
CartAggregateMappingExtensions.MapTo (PromotionEvaluationContext and
TaxEvaluationContext) and CartFilterMappingExtensions.MapTo were static
extension methods, which can't be overridden - the concrete failure mode
this caused was ToTaxAddress/ToProductPromoEntry's virtual seam only
working when called directly, not from the extensions' hard-coded static
calls. Moved all three onto IXCartMapper/XCartMapper as public virtual
methods; the calls that used to go through mapper.ToTaxAddress(...) /
mapper.ToProductPromoEntry(...) as a passed-in parameter are now plain
this-calls. Matches the already-merged precedent in profile-experience-api
and this repo's own ICatalogCsvImportModuleMapper, where in-place MapTo is
an interface method, and the same fix just made in x-catalog.

Threaded IXCartMapper through CartSearchCriteriaBuilder into
SearchCartQueryHandler/SearchWishlistQueryHandler, and through
MapPromoEvalContextMiddleware/LoadCartToEvalContextMiddleware for the two
CartAggregate overloads.

Also: criteria (the destination) can never legitimately be null in
MapTo(IList<IFilter>, ShoppingCartSearchCriteria) - it's always constructed
by the caller before this method fills it in - so a null criteria now
throws ArgumentNullException instead of silently no-oping. filters is still
allowed to be null (nothing to apply).
